requirements.txt
aiosqlite==0.17.0
anyio==3.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
black==21.12b0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
fastapi==0.70.1
flake8==4.0.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.1
hyperframe==6.0.1
idna==3.3
iso8601==0.1.16
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
priority==2.0.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pyflakes==2.4.0
pypika-tortoise==0.1.2
python-multipart==0.0.5
pytz==2021.3
rope==0.22.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
tomli==1.2.3
tortoise-orm==0.18.0
typing_extensions==4.0.1
wsproto==1.0.0
